feat(python): add remaining api functions
