+++ .github/workflows/ci-python.yml
@@ -1,0 +1,104 @@
+name: ci-python
+
+on:
+  pull_request:
+    branches:
+      - python-wrapper
+  push:
+    branches:
+      - python-wrapper
+
+jobs:
+  build_windows:
+    runs-on: windows-latest
+    steps:
+      - uses: actions/checkout@v1
+      - run: |
+          mkdir .\out\windows
+          cargo build --release
+          cp .\target\release\bbs.dll .\out\windows
+      # - run: yarn build:windows
+      - uses: actions/upload-artifact@v2
+        with:
+          name: windows
+          path: ./out/windows/*
+
+  build_linux:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v1
+      - run: yarn build:linux
+      - uses: actions/upload-artifact@v2
+        with:
+          name: linux
+          path: ./out/linux/*
+
+  build_macos:
+    runs-on: macos-latest
+    steps:
+      - uses: actions/checkout@v1
+      - run: yarn build:macos
+      - uses: actions/upload-artifact@v2
+        with:
+          name: macos
+          path: ./out/macos/darwin-x86_64/*
+
+
+  test_macos:
+    needs: [build_macos]
+    runs-on: macos-latest
+    strategy:
+      matrix:
+        python-version: [3.7, 3.8, 3.9]
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - uses: actions/download-artifact@v2
+        with:
+          name: macos
+          path: ./wrappers/python/ursa_bbs_signatures
+      - name: Run tests
+        run: python -m unittest discover
+        
+  test_linux:
+    needs: [build_linux]
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.7, 3.8, 3.9]
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - uses: actions/download-artifact@v2
+        with:
+          name: linux
+          path: ./wrappers/python/ursa_bbs_signatures
+      - name: Run tests
+        run: python -m unittest discover
+        
+  test_windows:
+    needs: [build_windows]
+    runs-on: windows-latest
+    strategy:
+      matrix:
+        python-version: [3.7, 3.8, 3.9]
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - uses: actions/download-artifact@v2
+        with:
+          name: windows
+          path: ./wrappers/python/ursa_bbs_signatures
+      - name: Run tests
+        run: python -m unittest discover
+
+
