+++ wrappers/python/ursa_bbs_signatures/_ffi/FfiException.py
@@ -1,0 +1,2 @@
+class FfiException(Exception):
+    pass
