--- wrappers/python/ursa_bbs_signatures/foreign_function_interface/FfiException.py
@@ -1,2 +1,0 @@
-class FfiException(Exception):
-    pass
