+++ wrappers/python/ursa_bbs_signatures/models/BbsException.py
@@ -1,5 +1,2 @@
-
-
-
 class BbsException(Exception):
-  pass
+    pass
