--- wrappers/python/ursa_bbs_signatures/models/__init__.py
@@ -1,3 +1,0 @@
-from .ProofMessage import ProofMessage, ProofMessageType
-
-__all__ = [ProofMessageType, ProofMessage]
