+++ wrappers/python/ursa_bbs_signatures/models/keys/BlindedBlsKeyPair.py
@@ -1,10 +1,10 @@
 from typing import Optional, Union
 
-from ursa_bbs_signatures.foreign_function_interface.bindings.bls import (
+from ..._ffi.bindings.bls import (
     bls_generate_blinded_g1_key,
-    bls_generate_blinded_g2_key,
+    bls_generate_blinded_g2_key, blinding_factor_size,
 )
-from ursa_bbs_signatures.foreign_function_interface.ffi_util import encode_bytes
+from ..._ffi.ffi_util import encode_bytes
 from .BlsKeyPair import BlsKeyPair
 
 
@@ -13,12 +13,14 @@ class BlindedBlsKeyPair(BlsKeyPair):
         super().__init__(public_key, secret_key=secret_key)
         self.blinding_factor = blinding_factor
 
+    @staticmethod
+    def blinding_factor_size() -> int:
+        return blinding_factor_size()
+
     @classmethod
     def generate_g1(
         cls, seed: Optional[Union[str, bytes]] = None
     ) -> "BlindedBlsKeyPair":
-        if seed:
-            seed = encode_bytes(seed)
         res = bls_generate_blinded_g1_key(seed)
         return cls(res["public_key"], res["secret_key"], res["blinding_factor"])
 
@@ -26,7 +28,5 @@ class BlindedBlsKeyPair(BlsKeyPair):
     def generate_g2(
         cls, seed: Optional[Union[str, bytes]] = None
     ) -> "BlindedBlsKeyPair":
-        if seed:
-            seed = encode_bytes(seed)
         res = bls_generate_blinded_g2_key(seed)
         return cls(res["public_key"], res["secret_key"], res["blinding_factor"])
