--- wrappers/python/ursa_bbs_signatures/models/keys/__init__.py
@@ -1,5 +1,0 @@
-from .BlindedBlsKeyPair import BlindedBlsKeyPair
-from .BbsKey import BbsKey
-from .BlsKeyPair import BlsKeyPair
-
-__all__ = [BbsKey, BlsKeyPair, BlindedBlsKeyPair]
